Remove prints and move payload
